Update prerequisite-layer.pkr.hcl
Introduction
This is a Packer configuration file that defines a prerequisite layer in creating a custom machine image in Azure. This configuration file creates an Azure resource that will be used as a source for the custom image. The configuration file is written in HashiCorp Configuration Language (HCL).

Variables
The configuration file defines the following variables:

subscription_id: This is the subscription ID for the Azure account where the resource group and managed image will be created.
client_id: This is the client ID for the Azure service principal that will be used to authenticate to the Azure API.
client_secret: This is the client secret for the Azure service principal that will be used to authenticate to the Azure API.
tenant_id: This is the tenant ID for the Azure account where the resource group and managed image will be created.
Azure Resource
The configuration file defines an Azure resource of type source.azure-arm.prerequisite_layer. This resource creates an Azure managed image that will be used as a source for the custom image.

The resource has the following properties:

subscription_id: This is the subscription ID for the Azure account where the resource group and managed image will be created. This is set to the subscription_id variable.
client_id: This is the client ID for the Azure service principal that will be used to authenticate to the Azure API. This is set to the client_id variable.
client_secret: This is the client secret for the Azure service principal that will be used to authenticate to the Azure API. This is set to the client_secret variable.
tenant_id: This is the tenant ID for the Azure account where the resource group and managed image will be created. This is set to the tenant_id variable.
managed_image_resource_group_name: This is the name of the resource group where the managed image will be created. This is set to "MainGroup".
managed_image_name: This is the name of the managed image that will be created. This is set to "windows-server-2019-with-dotnetcore".
os_type: This is the operating system type of the managed image. This is set to "Windows".
image_publisher: This is the publisher of the image used for the managed image. This is set to "MicrosoftWindowsServer".
image_offer: This is the offer of the image used for the managed image. This is set to "WindowsServer".
image_sku: This is the SKU of the image used for the managed image. This is set to "2019-Datacenter".
vm_size: This is the size of the virtual machine that will be used to create the managed image. This is set to "Standard_DS1_v2".
location: This is the location of the Azure region where the managed image will be created. This is set to "SwedenCentral".
communicator: This is the type of communicator that will be used to communicate with the virtual machine used to create the managed image. This is set to "winrm".
winrm_use_ssl: This indicates whether SSL should be used when communicating with the virtual machine. This is set to true.
winrm_insecure: This indicates whether the SSL certificate should be ignored when communicating with the virtual machine. This is set to true.
winrm_timeout: This is the maximum amount of time that Packer will wait for WinRM to become available on the virtual machine. This is set to "3m".
winrm_username: This is the username that will be used to authenticate to the virtual machine over WinRM. This is set to "packer".
winrm_password: This